@@ -10,83 +18,104 @@
   "bugs": {
     "url": "https://github.com/rapid7/awsaml/issues"
   },
-  "packageManager": "yarn@3.4.1",
+  "engines": {
+    "node": ">=16.0.0"
+  },
+  "packageManager": "yarn@3.5.0",
   "scripts": {
-    "electron": "electron electron/electron.js",
-    "electron-dev": "NODE_ENV=development ELECTRON_START_URL=http://localhost:3000 electron electron/electron.js",
-    "react-start": "BROWSER=none; NODE_ENV=development react-scripts start",
-    "react-build": "react-scripts build",
-    "test": "mocha",
-    "cover": "rm -Rf coverage/* && istanbul cover _mocha -- -R spec",
-    "lint": "eslint '*.js' 'electron/**/*.js' 'api/**/*.js' 'src/**/*.js' 'test/**/*.js'",
+    "electron": "electron src/main/index.js",
+    "electron-dev": "NODE_ENV=development ELECTRON_START_URL=http://localhost:3000 electron src/main/index.js",
+    "react-start": "BROWSER=none; NODE_ENV=development craco start",
+    "react-build": "craco build",
+    "test": "jest",
+    "lint": "eslint '*.js' 'src/**/*.js' 'test/**/*.js'",
     "report": "coveralls < ./coverage/lcov.info",
-    "prebuild": "rm -rf dist/ && rm -rf build/",
-    "build": "yarn react-build && node packager.js",
-    "postbuild": "for platform in `echo $PLATFORM | sed 's/,/ /g'`; do export platform=$platform; npm run zip; done",
-    "zip": "cd dist/Awsaml-${platform}-x64 && zip -q -y -FS -r ../awsaml-v${npm_package_version}-${platform}-x64.zip .",
-    "show-appcast-checkpoint": "curl --compressed --location --user-agent 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/33.0.1750.152 Safari/537.36' 'https://github.com/rapid7/awsaml/releases.atom' | /usr/bin/sed 's|<pubDate>[^<]*</pubDate>||g' | shasum --algorithm 256"
+    "build": "node build.js",
+    "show-appcast-checkpoint": "curl --compressed --location --user-agent 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/33.0.1750.152 Safari/537.36' 'https://github.com/rapid7/awsaml/releases.atom' | /usr/bin/sed 's|<pubDate>[^<]*</pubDate>||g' | shasum --algorithm 256",
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make",
+    "publish": "electron-forge publish",
+    "clean": "rm -rf out && rm -rf build"
   },
   "homepage": "./",
   "proxy": "http://localhost:2600/",
-  "main": "electron/electron.js",
+  "main": "src/main/index.js",
   "dependencies": {
-    "aws-sdk": "^2.1094.0",
-    "body-parser": "^1.19.2",
+    "@aws-sdk/client-sts": "^3.370.0",
+    "@node-saml/passport-saml": "^4.0.4",
+    "@xmldom/xmldom": "^0.8.9",
+    "body-parser": "^1.20.2",
     "cookie-parser": "^1.4.6",
-    "errorhandler": "^1.5.0",
-    "express": "^4.17.3",
-    "express-session": "^1.17.2",
-    "ini": "^1.3.8",
-    "lodash.isplainobject": "^4.0.6",
-    "mkdirp": "^0.5.5",
+    "electron-squirrel-startup": "^1.0.0",
+    "express": "^4.18.2",
+    "express-session": "^1.17.3",
+    "ini": "^4.1.1",
     "morgan": "^1.10.0",
-    "passport": "^0.4.1",
-    "passport-saml": "^3.1.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "uuid": "^3.4.0",
-    "xmldom": "^0.6.0",
+    "passport": "^0.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "stylis": "^4.3.0",
+    "update-electron-app": "^2.0.1",
+    "uuid": "^9.0.0",
     "xpath.js": "^1.1.0"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome": "^1.1.8",
-    "@fortawesome/fontawesome-free-regular": "^5.0.13",
-    "@fortawesome/fontawesome-free-solid": "^5.0.13",
-    "@fortawesome/react-fontawesome": "^0.0.20",
-    "axios": "^0.21.2",
-    "babel-eslint": "^8.2.5",
-    "bootstrap": "^4.6.1",
+    "@babel/core": "^7.22.9",
+    "@babel/eslint-parser": "^7.22.9",
+    "@babel/preset-env": "^7.22.9",
+    "@babel/preset-react": "^7.22.5",
+    "@craco/craco": "^7.1.0",
+    "@electron-forge/cli": "^6.2.1",
+    "@electron-forge/core": "^6.2.1",
+    "@electron-forge/maker-deb": "^6.2.1",
+    "@electron-forge/maker-rpm": "^6.2.1",
+    "@electron-forge/maker-squirrel": "^6.2.1",
+    "@electron-forge/maker-zip": "^6.2.1",
+    "@electron-forge/plugin-webpack": "^6.2.1",
+    "@electron-forge/publisher-github": "^6.2.1",
+    "@electron/get": "^2.0.2",
+    "@electron/rebuild": "^3.2.13",
+    "@fortawesome/fontawesome-free": "^6.4.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-brands-svg-icons": "^6.4.0",
+    "@fortawesome/free-regular-svg-icons": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@popperjs/core": "^2.11.8",
+    "babel-jest": "^29.6.1",
+    "bootstrap": "^5.3.0",
     "coveralls": "^3.1.1",
-    "electron": "^11.5.0",
-    "electron-devtools-installer": "^2.2.4",
-    "electron-packager": "^15.4.0",
-    "electron-rebuild": "^1.11.0",
-    "eslint": "^5.0.1",
-    "eslint-config-rapid7": "^3.0.2",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-react": "^7.29.4",
-    "history": "^4.10.1",
-    "hoist-non-react-statics": "^2.5.5",
-    "istanbul": "^0.4.5",
-    "jquery": "^3.6.0",
-    "mocha": "^5.2.0",
-    "mocha-lcov-reporter": "^1.3.0",
-    "popper.js": "^1.16.1",
-    "prismjs": "^1.27.0",
+    "electron": "^25.3.0",
+    "electron-packager": "^17.1.1",
+    "eslint": "^8.45.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.3",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "glob": "^10.3.3",
+    "history": "^5.3.0",
+    "jest": "^29.6.1",
+    "jest-junit": "^16.0.0",
+    "prismjs": "^1.29.0",
     "prop-types": "^15.8.1",
-    "querystring": "^0.2.1",
-    "react-clipboard.js": "^2.0.16",
-    "react-redux": "^5.1.2",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-router-redux": "^5.0.0-alpha.9",
-    "react-scripts": "^1.1.4",
-    "reactstrap": "^6.1.0",
-    "redux": "^4.1.2",
-    "redux-actions": "^2.4.0",
-    "redux-thunk": "^2.4.1",
-    "should": "^13.2.1",
-    "styled-components": "^3.3.3",
-    "unchanged": "^1.3.1"
+    "react-is": "^18.2.0",
+    "react-router": "^6.14.1",
+    "react-router-dom": "^6.14.1",
+    "react-scripts": "^5.0.1",
+    "reactstrap": "^9.2.0",
+    "should": "^13.2.3",
+    "styled-components": "^6.0.4"
+  },
+  "browserslist": {
+    "production": [
+      "last 1 electron version"
+    ],
+    "development": [
+      "last 1 electron version"
+    ]
   }
 }
